moved ui component files to src/components/
